requirements.txt
boto3==1.20.26
requests==2.26.0
lz4==3.1.10
Pillow==8.4.0